Fix fantom require cycle
Resolves the warning

Require cycle: src/plugins/stake-plugins/uniswapV2/policyInfo/fantom.ts -> src/plugins/stake-plugins/uniswapV2/policyInfo/fantom.ts

Require cycles are allowed, but can result in uninitialized values. Consider refactoring to remove the need for a cycle.
